+++ src/00_core.lua
@@ -34,7 +34,7 @@
 
 _G.Ess = _G.Ess or {}
 local Ess = _G.Ess
-Ess.VERSION = "0.5.0"
+Ess.VERSION = "0.5.1"
 
 Ess.Safe = Ess.Safe or {}
 Ess.Table = Ess.Table or {}